Formatting & Documentation
Formatting of the rspec files to meet `pdk validate`'s requirements.
Also includes the addition/correction of the appropriate YARD-tags to
generate a REFERENCE.md file. Not all components have been updated to
match this spec, however, most do. Some of the Documentation could use a
bit of tlc to make it clearer and easier to extend.